+++ fineract-accounting/src/main/java/org/apache/fineract/accounting/journalentry/data/AccountingBridgeDataDTO.java
@@ -16,41 +16,39 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.fineract.portfolio.loanaccount.data;
+package org.apache.fineract.accounting.journalentry.data;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AccountingBridgeLoanTransactionDTO {
+public class AccountingBridgeDataDTO {
 
-    private Long id;
+    private Long loanId;
+    private Long loanProductId;
     private Long officeId;
-    private LoanTransactionEnumData type;
-    private boolean reversed;
-    private LocalDate date;
     private String currencyCode;
-    private BigDecimal amount;
-    private BigDecimal netDisbursalAmount;
-    private BigDecimal principalPortion;
-    private BigDecimal interestPortion;
-    private BigDecimal feeChargesPortion;
-    private BigDecimal penaltyChargesPortion;
-    private BigDecimal overPaymentPortion;
-    private String chargeRefundChargeType;
-    private Long paymentTypeId;
-    private List<LoanChargePaidByDTO> loanChargesPaid = new ArrayList<>();
-    private BigDecimal principalPaid;
-    private BigDecimal feePaid;
-    private BigDecimal penaltyPaid;
-    private LoanChargeData loanChargeData;
-    private boolean loanToLoanTransfer;
+    private BigDecimal calculatedInterest;
+    private boolean cashBasedAccountingEnabled;
+    private boolean upfrontAccrualBasedAccountingEnabled;
+    private boolean periodicAccrualBasedAccountingEnabled;
+    private boolean isAccountTransfer;
+    private boolean isChargeOff;
+    private boolean isFraud;
+    private Long chargeOffReasonCodeValue;
+    private boolean isWrittenOff;
+    private List<AccountingBridgeLoanTransactionDTO> newLoanTransactions = new ArrayList<>();
+    private boolean merchantBuyDownFee;
+    private List<AdvancedMappingtDTO> buydownFeeClassificationCodeValue;
+    private List<AdvancedMappingtDTO> capitalizedIncomeClassificationCodeValue;
+    private AdvancedMappingtDTO writeOffReasonCodeValue;
 
 }
