+++ fineract-accounting/src/main/java/org/apache/fineract/accounting/journalentry/data/LoanChargePaidByDTO.java
@@ -16,25 +16,25 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.fineract.portfolio.loanaccount.data;
+package org.apache.fineract.accounting.journalentry.data;
 
 import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/**
- * Carries the pro-rated tax amount for a single TaxComponent when a LoanCharge is (partially) paid. Used to propagate
- * tax details from the domain layer to the accounting bridge.
- */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChargeTaxDetailDTO {
+public class LoanChargePaidByDTO {
 
-    /** GL account to credit (tax liability account from TaxComponent.creditAccount). */
-    private Long creditAccountId;
-
-    /** Pro-rated tax amount for this component in this payment. */
+    private Long chargeId;
+    private Boolean isPenalty;
+    private Long loanChargeId;
     private BigDecimal amount;
+    private Integer installmentNumber;
+    private List<ChargeTaxDetailDTO> taxDetails = new ArrayList<>();
+
 }
