+++ fineract-accounting/src/main/java/org/apache/fineract/accounting/journalentry/service/JournalEntryWritePlatformService.java
@@ -21,14 +21,9 @@ package org.apache.fineract.accounting.journalentry.service;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Map;
-import org.apache.fineract.accounting.provisioning.domain.ProvisioningEntry;
+import org.apache.fineract.accounting.journalentry.data.AccountingBridgeDataDTO;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-import org.apache.fineract.investor.domain.ExternalAssetOwner;
-import org.apache.fineract.investor.domain.ExternalAssetOwnerTransfer;
-import org.apache.fineract.portfolio.loanaccount.data.AccountingBridgeDataDTO;
-import org.apache.fineract.portfolio.loanaccount.domain.Loan;
-import org.apache.fineract.portfolio.loanaccount.domain.LoanTransaction;
 
 public interface JournalEntryWritePlatformService {
 
@@ -46,37 +41,9 @@ public interface JournalEntryWritePlatformService {
 
     void createJournalEntryForReversedLoanTransaction(LocalDate transactionDate, String loanTransactionId, Long officeId);
 
-    String revertProvisioningJournalEntries(LocalDate reversalTransactionDate, Long entityId, Integer entityType);
-
-    String createProvisioningJournalEntries(ProvisioningEntry entry);
-
     void createJournalEntriesForShares(Map<String, Object> accountingBridgeData);
 
     void revertShareAccountJournalEntries(ArrayList<Long> transactionId, LocalDate transactionDate);
 
-    /**
-     * Create journal entries immediately for a single loan transaction
-     *
-     * @param loanTransaction
-     *            the loan transaction to create journal entries for
-     * @param isAccountTransfer
-     *            whether this is an account transfer transaction
-     * @param isLoanToLoanTransfer
-     *            whether this is a loan-to-loan transfer transaction
-     */
-    void createJournalEntriesForLoanTransaction(LoanTransaction loanTransaction, boolean isAccountTransfer, boolean isLoanToLoanTransfer);
-
-    /**
-     * Create journal entries immediately for an external owner transfer
-     *
-     * @param loan
-     *            the loan being transferred
-     * @param externalAssetOwnerTransfer
-     *            the external owner transfer details
-     * @param previousOwner
-     *            the previous owner (can be null for initial transfers)
-     */
-    void createJournalEntriesForExternalOwnerTransfer(Loan loan, ExternalAssetOwnerTransfer externalAssetOwnerTransfer,
-            ExternalAssetOwner previousOwner);
-
+    void createJournalEntriesForLoanTransaction(AccountingBridgeDataDTO accountingBridgeDataDTO, boolean isLoanToLoanTransfer);
 }
