+++ fineract-provider/src/main/java/org/apache/fineract/accounting/journalentry/service/CashBasedAccountingProcessorForLoan.java
@@ -35,9 +35,9 @@ import org.apache.fineract.accounting.journalentry.data.ChargeTaxPaymentDTO;
 import org.apache.fineract.accounting.journalentry.data.GLAccountBalanceHolder;
 import org.apache.fineract.accounting.journalentry.data.LoanDTO;
 import org.apache.fineract.accounting.journalentry.data.LoanTransactionDTO;
+import org.apache.fineract.accounting.journalentry.data.LoanTransactionTypeDTO;
 import org.apache.fineract.infrastructure.core.service.MathUtil;
 import org.apache.fineract.organisation.office.domain.Office;
-import org.apache.fineract.portfolio.loanaccount.data.LoanTransactionEnumData;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -62,7 +62,7 @@ public class CashBasedAccountingProcessorForLoan implements AccountingProcessorF
             final String transactionId = loanTransactionDTO.getTransactionId();
             final Long paymentTypeId = loanTransactionDTO.getPaymentTypeId();
             final Long loanId = loanDTO.getLoanId();
-            final LoanTransactionEnumData transactionType = loanTransactionDTO.getTransactionType();
+            final LoanTransactionTypeDTO transactionType = loanTransactionDTO.getTransactionType();
 
             this.helper.checkForBranchClosures(latestGLClosure, transactionDate);
 
@@ -301,9 +301,9 @@ public class CashBasedAccountingProcessorForLoan implements AccountingProcessorF
         }
 
         if (totalDebitAmount.compareTo(BigDecimal.ZERO) > 0) {
-            Long chargeId = loanTransactionDTO.getLoanChargeData().getChargeId();
+            Long chargeId = loanTransactionDTO.getLoanChargeDTO().getChargeId();
             Integer accountMappingTypeId;
-            if (loanTransactionDTO.getLoanChargeData().isPenalty()) {
+            if (loanTransactionDTO.getLoanChargeDTO().isPenalty()) {
                 accountMappingTypeId = CashAccountsForLoan.INCOME_FROM_PENALTIES.getValue();
             } else {
                 accountMappingTypeId = CashAccountsForLoan.INCOME_FROM_FEES.getValue();
@@ -399,9 +399,9 @@ public class CashBasedAccountingProcessorForLoan implements AccountingProcessorF
         }
 
         if (totalDebitAmount.compareTo(BigDecimal.ZERO) > 0) {
-            Long chargeId = loanTransactionDTO.getLoanChargeData().getChargeId();
+            Long chargeId = loanTransactionDTO.getLoanChargeDTO().getChargeId();
             Integer accountMappingTypeId;
-            if (loanTransactionDTO.getLoanChargeData().isPenalty()) {
+            if (loanTransactionDTO.getLoanChargeDTO().isPenalty()) {
                 accountMappingTypeId = CashAccountsForLoan.INCOME_FROM_PENALTIES.getValue();
             } else {
                 accountMappingTypeId = CashAccountsForLoan.INCOME_FROM_FEES.getValue();
