+++ fineract-provider/src/test/java/org/apache/fineract/accounting/journalentry/CreateJournalEntriesForChargeOffLoanTest.java
@@ -33,7 +33,7 @@ import org.apache.fineract.accounting.common.AccountingConstants.AccrualAccounts
 import org.apache.fineract.accounting.glaccount.domain.GLAccount;
 import org.apache.fineract.accounting.journalentry.data.LoanDTO;
 import org.apache.fineract.accounting.journalentry.data.LoanTransactionDTO;
-import org.apache.fineract.accounting.journalentry.service.AccountingProcessorHelper;
+import org.apache.fineract.accounting.journalentry.service.AccountingProcessorHelperImpl;
 import org.apache.fineract.accounting.journalentry.service.AccrualBasedAccountingProcessorForLoan;
 import org.apache.fineract.accounting.producttoaccountmapping.domain.ProductToGLAccountMapping;
 import org.apache.fineract.organisation.office.domain.Office;
@@ -52,7 +52,7 @@ class CreateJournalEntriesForChargeOffLoanTest {
     private static final Long chargeOffReasonId = 15L;
 
     @Mock
-    private AccountingProcessorHelper helper;
+    private AccountingProcessorHelperImpl helper;
     @InjectMocks
     private AccrualBasedAccountingProcessorForLoan processor;
     private LoanDTO loanDTO;
