+++ fineract-provider/src/test/java/org/apache/fineract/accounting/journalentry/CreateJournalEntriesForTransferLoanTest.java
@@ -34,7 +34,7 @@ import org.apache.fineract.accounting.closure.domain.GLClosure;
 import org.apache.fineract.accounting.common.AccountingConstants.AccrualAccountsForLoan;
 import org.apache.fineract.accounting.journalentry.data.LoanDTO;
 import org.apache.fineract.accounting.journalentry.data.LoanTransactionDTO;
-import org.apache.fineract.accounting.journalentry.service.AccountingProcessorHelper;
+import org.apache.fineract.accounting.journalentry.service.AccountingProcessorHelperImpl;
 import org.apache.fineract.accounting.journalentry.service.AccrualBasedAccountingProcessorForLoan;
 import org.apache.fineract.organisation.office.domain.Office;
 import org.apache.fineract.portfolio.loanaccount.data.LoanTransactionEnumData;
@@ -59,7 +59,7 @@ class CreateJournalEntriesForTransferLoanTest {
     private static final BigDecimal PRINCIPAL_AMOUNT = new BigDecimal("500.00");
 
     @Mock
-    private AccountingProcessorHelper helper;
+    private AccountingProcessorHelperImpl helper;
     @InjectMocks
     private AccrualBasedAccountingProcessorForLoan processor;
     private Office office;
