+++ fineract-savings/src/main/java/org/apache/fineract/portfolio/savings/service/DepositAccountWritePlatformService.java
@@ -18,11 +18,8 @@
  */
 package org.apache.fineract.portfolio.savings.service;
 
-import java.time.LocalDate;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-import org.apache.fineract.organisation.office.domain.Office;
-import org.apache.fineract.organisation.staff.domain.Staff;
 import org.apache.fineract.portfolio.savings.DepositAccountType;
 import org.apache.fineract.portfolio.savings.data.SavingsAccountTransactionDTO;
 import org.apache.fineract.portfolio.savings.domain.SavingsAccountTransaction;
@@ -61,28 +58,10 @@ public interface DepositAccountWritePlatformService {
 
     CommandProcessingResult prematureCloseRDAccount(Long savingsId, JsonCommand command);
 
-    SavingsAccountTransaction initiateSavingsTransfer(Long accountId, LocalDate transferDate, DepositAccountType depositAccountType);
-
-    SavingsAccountTransaction withdrawSavingsTransfer(Long accountId, LocalDate transferDate, DepositAccountType depositAccountType);
-
-    void rejectSavingsTransfer(Long accountId, DepositAccountType depositAccountType);
-
-    SavingsAccountTransaction acceptSavingsTransfer(Long accountId, LocalDate transferDate, Office acceptedInOffice, Staff staff,
-            DepositAccountType depositAccountType);
-
-    CommandProcessingResult addSavingsAccountCharge(JsonCommand command, DepositAccountType depositAccountType);
-
-    CommandProcessingResult updateSavingsAccountCharge(JsonCommand command, DepositAccountType depositAccountType);
-
-    CommandProcessingResult deleteSavingsAccountCharge(Long savingsAccountId, Long savingsAccountChargeId, JsonCommand command,
-            DepositAccountType depositAccountType);
-
     CommandProcessingResult waiveCharge(Long savingsAccountId, Long savingsAccountChargeId, DepositAccountType depositAccountType);
 
     CommandProcessingResult payCharge(Long savingsAccountId, Long savingsAccountChargeId, JsonCommand command,
             DepositAccountType depositAccountType);
-
-    void applyChargeDue(Long savingsAccountChargeId, Long accountId, DepositAccountType depositAccountType);
 
     void updateMaturityDetails(Long depositAccountId, DepositAccountType depositAccountType);
